feat: extract fact materializer
