+++ packages/server/src/entities/recreate.ts
@@ -1,15 +1,14 @@
-import type { Kysely, Selectable } from "kysely";
+import type { Kysely } from "kysely";
 import { sql } from "kysely";
 import type { Logger } from "pino";
 import { createEmbeddingProvider } from "../connectors/embeddings";
 import { type EnrichmentResult, isEnrichmentActive, runEnrichment } from "../connectors/enrichment";
-import { normalizeName } from "../connectors/name-normalize";
 import { getSyncProgress } from "../connectors/sync";
-import { createEntityRepository } from "../db/repositories/entities";
-import { createEntityReviewRepo } from "../db/repositories/entity-review";
-import type { DB, EntitiesTable } from "../db/schema";
-import { type Entity, type EntityLookup, type ProposeEntityType, proposeEntity } from "./propose";
+import type { DB } from "../db/schema";
+import { type ReplayFactsSummary, replaySourceFacts } from "./materialize";
 import { isRecreateActive, withRecreateLock } from "./recreate-state";
+
+export type { ReplayFactsSummary } from "./materialize";
 
 export interface ResetSummary {
   dryRun: boolean;
@@ -170,413 +169,6 @@ export async function resetDerivedEntityData(
 }
 
 // ─────────────────────────────────────────────────────────────────────────
-// Phase 2 — Replay durable source facts back into the entity graph.
-// ─────────────────────────────────────────────────────────────────────────
-
-export interface ReplayFactsSummary {
-  factsRead: number;
-  entitiesCreated: number;
-  entitiesLinked: number;
-  queued: number;
-  mentionsWritten: number;
-  skipped: number;
-}
-
-const FACT_REPLAY_ORDER = [
-  "structural_seed",
-  "person_seed",
-  "attendee",
-  "assignee",
-  "author",
-  "parent_entity",
-] as const;
-
-type EntityRow = Selectable<EntitiesTable>;
-
-function readJsonObject(raw: string | null): Record<string, unknown> {
-  if (!raw) return {};
-  try {
-    const parsed = JSON.parse(raw);
-    return parsed && typeof parsed === "object" ? (parsed as Record<string, unknown>) : {};
-  } catch {
-    return {};
-  }
-}
-
-function readPersonEmailFromMetadata(metadata: string | null): string | null {
-  if (!metadata) return null;
-  try {
-    const parsed = JSON.parse(metadata) as { email?: unknown };
-    if (typeof parsed.email === "string" && parsed.email.length > 0) return parsed.email.toLowerCase();
-  } catch {
-    /* corrupt metadata — skip */
-  }
-  return null;
-}
-
-function parseAliasesString(aliases: string | null): string[] {
-  if (!aliases) return [];
-  try {
-    const parsed = JSON.parse(aliases);
-    if (Array.isArray(parsed)) return parsed.filter((v): v is string => typeof v === "string");
-  } catch {
-    /* corrupt aliases — skip */
-  }
-  return [];
-}
-
-interface LookupIndex {
-  personEntities: EntityRow[];
-  byNormalizedName: Map<string, EntityRow[]>;
-  byNormalizedAlias: Map<string, EntityRow[]>;
-  bySourceRef: Map<string, EntityRow>;
-}
-
-async function buildLookupIndex(db: Kysely<DB>): Promise<LookupIndex> {
-  const personEntities = await db.selectFrom("entities").selectAll().where("source_type", "=", "person").execute();
-  const byNormalizedName = new Map<string, EntityRow[]>();
-  const byNormalizedAlias = new Map<string, EntityRow[]>();
-  for (const e of personEntities) {
-    const nameKey = normalizeName(e.name);
-    if (nameKey) {
-      const bucket = byNormalizedName.get(nameKey);
-      if (bucket) bucket.push(e);
-      else byNormalizedName.set(nameKey, [e]);
-    }
-    for (const alias of parseAliasesString(e.aliases)) {
-      const aliasKey = normalizeName(alias);
-      if (!aliasKey) continue;
-      const bucket = byNormalizedAlias.get(aliasKey);
-      if (bucket) bucket.push(e);
-      else byNormalizedAlias.set(aliasKey, [e]);
-    }
-  }
-  const sourceRefs = await db
-    .selectFrom("entity_source_refs")
-    .innerJoin("entities", "entities.id", "entity_source_refs.entity_id")
-    .select(["entity_source_refs.source as source", "entity_source_refs.source_id as source_id"])
-    .selectAll("entities")
-    .execute();
-  const bySourceRef = new Map<string, EntityRow>();
-  for (const row of sourceRefs) {
-    bySourceRef.set(`${row.source}:${row.source_id}`, row as unknown as EntityRow);
-  }
-  return { personEntities, byNormalizedName, byNormalizedAlias, bySourceRef };
-}
-
-function registerPerson(index: LookupIndex, entity: EntityRow): void {
-  if (!index.personEntities.some((p) => p.id === entity.id)) index.personEntities.push(entity);
-  const nameKey = normalizeName(entity.name);
-  if (nameKey) {
-    const bucket = index.byNormalizedName.get(nameKey);
-    if (bucket) {
-      if (!bucket.some((b) => b.id === entity.id)) bucket.push(entity);
-    } else {
-      index.byNormalizedName.set(nameKey, [entity]);
-    }
-  }
-  for (const alias of parseAliasesString(entity.aliases)) {
-    const aliasKey = normalizeName(alias);
-    if (!aliasKey) continue;
-    const bucket = index.byNormalizedAlias.get(aliasKey);
-    if (bucket) {
-      if (!bucket.some((b) => b.id === entity.id)) bucket.push(entity);
-    } else {
-      index.byNormalizedAlias.set(aliasKey, [entity]);
-    }
-  }
-}
-
-interface FactRow {
-  id: string;
-  indexed_file_id: string | null;
-  source: string;
-  fact_type: string;
-  relation: string;
-  subject_name: string | null;
-  subject_email: string | null;
-  subject_source: string | null;
-  subject_source_id: string | null;
-  context_snippet: string | null;
-  raw: string | null;
-}
-
-export async function replaySourceFacts(
-  db: Kysely<DB>,
-  logger: Logger,
-  opts: { triggeredByUserId: string },
-): Promise<ReplayFactsSummary> {
-  const summary: ReplayFactsSummary = {
-    factsRead: 0,
-    entitiesCreated: 0,
-    entitiesLinked: 0,
-    queued: 0,
-    mentionsWritten: 0,
-    skipped: 0,
-  };
-
-  const entityRepo = createEntityRepository(db);
-  const reviewRepo = createEntityReviewRepo(db);
-  const index = await buildLookupIndex(db);
-
-  const lookup: EntityLookup = {
-    getByNormalizedName: (n) => index.byNormalizedName.get(n) ?? [],
-    getByAlias: (n) => index.byNormalizedAlias.get(n) ?? [],
-    listByType: (t: ProposeEntityType) => (t === "person" ? index.personEntities : []),
-  };
-
-  const proposeDeps = {
-    entityRepo,
-    reviewRepo,
-    lookup,
-    readEmail: (e: Entity) => readPersonEmailFromMetadata(e.metadata),
-  };
-
-  // Cache file→connector_config_id mapping so personal/structural seeds can
-  // attribute correctly. Replay-time `triggeredByUserId` defaults to the caller
-  // but falls back to the connector_config.created_by when a fact has a file.
-  const fileToConnector = new Map<string, string>();
-  const connectorOwners = new Map<string, string>();
-  const allFiles = await db.selectFrom("indexed_files").select(["id", "connector_config_id"]).execute();
-  for (const f of allFiles) fileToConnector.set(f.id, f.connector_config_id);
-  const allConfigs = await db.selectFrom("connector_configs").select(["id", "created_by"]).execute();
-  for (const c of allConfigs) connectorOwners.set(c.id, c.created_by);
-
-  function resolveTriggeredBy(indexedFileId: string | null): string {
-    if (indexedFileId) {
-      const cfg = fileToConnector.get(indexedFileId);
-      if (cfg) {
-        const owner = connectorOwners.get(cfg);
-        if (owner) return owner;
-      }
-    }
-    return opts.triggeredByUserId;
-  }
-
-  // Pre-load all facts ordered by fact_type so the in-memory loop respects the
-  // fixed replay order (structural seeds → person seeds → people → parent refs).
-  const orderRank = new Map<string, number>(FACT_REPLAY_ORDER.map((t, i) => [t, i]));
-  const facts: FactRow[] = (
-    await db
-      .selectFrom("indexed_file_facts")
-      .select([
-        "id",
-        "indexed_file_id",
-        "source",
-        "fact_type",
-        "relation",
-        "subject_name",
-        "subject_email",
-        "subject_source",
-        "subject_source_id",
-        "context_snippet",
-        "raw",
-      ])
-      .execute()
-  )
-    .filter((f) => orderRank.has(f.fact_type))
-    .sort((a, b) => (orderRank.get(a.fact_type) ?? 0) - (orderRank.get(b.fact_type) ?? 0));
-
-  summary.factsRead = facts.length;
-
-  for (const fact of facts) {
-    try {
-      if (fact.fact_type === "structural_seed") {
-        await replayStructuralSeed(db, entityRepo, index, fact);
-      } else if (fact.fact_type === "person_seed") {
-        await replayPersonSeed(entityRepo, index, fact);
-      } else if (fact.fact_type === "attendee" || fact.fact_type === "assignee" || fact.fact_type === "author") {
-        await replayPersonFact(entityRepo, proposeDeps, index, fact, resolveTriggeredBy(fact.indexed_file_id), summary);
-      } else if (fact.fact_type === "parent_entity") {
-        await replayParentEntity(db, entityRepo, index, fact, summary);
-      } else {
-        summary.skipped++;
-      }
-    } catch (err) {
-      logger.warn({ err, factId: fact.id, factType: fact.fact_type }, "Replay failed for fact");
-      summary.skipped++;
-    }
-  }
-
-  logger.info({ summary }, "Source-fact replay complete");
-  return summary;
-}
-
-async function replayStructuralSeed(
-  db: Kysely<DB>,
-  entityRepo: ReturnType<typeof createEntityRepository>,
-  index: LookupIndex,
-  fact: FactRow,
-): Promise<void> {
-  const raw = readJsonObject(fact.raw);
-  const subjectSource = fact.subject_source ?? fact.source;
-  const subjectSourceId = fact.subject_source_id;
-  if (!subjectSourceId || !fact.subject_name) return;
-
-  // File-attached structural seed (promotable file → structural entity):
-  // sourceType is `${connectorType}_${fileType}`, sourceRefId points at the
-  // indexed file. File-less structural seed (onEntitySeed payload): use raw
-  // seed shape — sourceType is provided by the caller of onEntitySeed.
-  let sourceType: string;
-  if (typeof raw.sourceType === "string") {
-    sourceType = raw.sourceType;
-  } else if (typeof raw.fileType === "string") {
-    sourceType = `${fact.source}_${raw.fileType}`;
-  } else {
-    sourceType = fact.source;
-  }
-  const sourceUrl = typeof raw.providerUrl === "string" ? raw.providerUrl : (raw.sourceUrl as string | undefined);
-  const sourcePath = typeof raw.sourcePath === "string" ? raw.sourcePath : undefined;
-  const metadataFromRaw =
-    raw.metadata && typeof raw.metadata === "object" ? (raw.metadata as Record<string, unknown>) : undefined;
-  const metadata = metadataFromRaw ?? (sourcePath ? { path: sourcePath } : undefined);
-
-  const entity = await entityRepo.upsertEntityFromTool({
-    name: fact.subject_name,
-    sourceType,
-    source: subjectSource,
-    sourceId: subjectSourceId,
-    sourceUrl,
-    sourceRefId: fact.indexed_file_id ?? undefined,
-    metadata,
-  });
-  index.bySourceRef.set(`${subjectSource}:${subjectSourceId}`, entity as unknown as EntityRow);
-}
-
-async function replayPersonSeed(
-  entityRepo: ReturnType<typeof createEntityRepository>,
-  index: LookupIndex,
-  fact: FactRow,
-): Promise<void> {
-  if (!fact.subject_name || !fact.subject_source || !fact.subject_source_id) return;
-  const raw = readJsonObject(fact.raw);
-  const subtype = raw.subtype === "internal" ? "internal" : "external";
-  const entity = await entityRepo.upsertPersonEntity({
-    name: fact.subject_name,
-    email: fact.subject_email ?? undefined,
-    subtype,
-    source: fact.subject_source,
-    sourceId: fact.subject_source_id,
-  });
-  index.bySourceRef.set(`${fact.subject_source}:${fact.subject_source_id}`, entity as unknown as EntityRow);
-  registerPerson(index, entity as unknown as EntityRow);
-}
-
-const PERSON_FACT_RELATION = {
-  attendee: "attended",
-  assignee: "assigned",
-  author: "authored",
-} as const;
-
-async function replayPersonFact(
-  entityRepo: ReturnType<typeof createEntityRepository>,
-  proposeDeps: {
-    entityRepo: ReturnType<typeof createEntityRepository>;
-    reviewRepo: ReturnType<typeof createEntityReviewRepo>;
-    lookup: EntityLookup;
-    readEmail: (e: Entity) => string | null;
-  },
-  index: LookupIndex,
-  fact: FactRow,
-  triggeredByUserId: string,
-  summary: ReplayFactsSummary,
-): Promise<void> {
-  if (!fact.subject_name) {
-    summary.skipped++;
-    return;
-  }
-  const factType = fact.fact_type as keyof typeof PERSON_FACT_RELATION;
-  const relation = PERSON_FACT_RELATION[factType];
-  const subtype = fact.subject_email ? "external" : "external";
-
-  let entity: EntityRow | null = null;
-  if (fact.subject_source && fact.subject_source_id) {
-    const refKey = `${fact.subject_source}:${fact.subject_source_id}`;
-    const cached = index.bySourceRef.get(refKey);
-    if (cached) {
-      entity = cached;
-    }
-  }
-
-  if (!entity) {
-    const source = fact.subject_source ?? fact.source;
-    const sourceId =
-      fact.subject_source_id ?? `${fact.indexed_file_id ?? "no-file"}:${fact.subject_email ?? fact.subject_name}`;
-    const result = await proposeEntity(proposeDeps, {
-      name: fact.subject_name,
-      email: fact.subject_email ?? null,
-      entityType: "person",
-      subtype,
-      source,
-      sourceId,
-      evidence: fact.indexed_file_id ? [{ indexedFileId: fact.indexed_file_id }] : [],
-      triggeredByUserId,
-    });
-    if (result.kind === "queued") {
-      summary.queued++;
-      return;
-    }
-    entity = result.entity as unknown as EntityRow;
-    if (result.kind === "created") summary.entitiesCreated++;
-    if (result.kind === "linked") summary.entitiesLinked++;
-    registerPerson(index, entity);
-    if (fact.subject_source && fact.subject_source_id) {
-      index.bySourceRef.set(`${fact.subject_source}:${fact.subject_source_id}`, entity);
-    }
-  }
-
-  if (!entity || !fact.indexed_file_id) return;
-  await entityRepo.createMention({
-    entityId: entity.id,
-    indexedFileId: fact.indexed_file_id,
-    contextSnippet: fact.context_snippet ?? null,
-    confidence: "EXTRACTED",
-    source: `${fact.source}_${fact.fact_type}`,
-    relation,
-  });
-  summary.mentionsWritten++;
-}
-
-async function replayParentEntity(
-  db: Kysely<DB>,
-  entityRepo: ReturnType<typeof createEntityRepository>,
-  index: LookupIndex,
-  fact: FactRow,
-  summary: ReplayFactsSummary,
-): Promise<void> {
-  if (!fact.indexed_file_id || !fact.subject_source || !fact.subject_source_id) {
-    summary.skipped++;
-    return;
-  }
-  const refKey = `${fact.subject_source}:${fact.subject_source_id}`;
-  let entity = index.bySourceRef.get(refKey);
-  if (!entity) {
-    const found = await entityRepo.getEntityBySourceRef(fact.subject_source, fact.subject_source_id);
-    if (found) {
-      entity = found as unknown as EntityRow;
-      index.bySourceRef.set(refKey, entity);
-    }
-  }
-  if (!entity) {
-    // The parent structural seed didn't replay — likely because the
-    // structural_seed fact never landed for it. Skip rather than fabricate
-    // a fresh entity from a parent-only reference (which would create
-    // duplicates if the seed lands later).
-    summary.skipped++;
-    return;
-  }
-  await entityRepo.createMention({
-    entityId: entity.id,
-    indexedFileId: fact.indexed_file_id,
-    contextSnippet: fact.context_snippet ?? null,
-    confidence: "EXTRACTED",
-    source: `${fact.source}_parent_entity`,
-    relation: "mentioned",
-  });
-  summary.mentionsWritten++;
-}
-
-// ─────────────────────────────────────────────────────────────────────────
 // Phase 3 — Reset → replay → enrichment loop.
 // ─────────────────────────────────────────────────────────────────────────
 
@@ -645,7 +237,7 @@ export async function recreateEntityGraph(deps: RecreateDeps): Promise<RecreateS
         }
       : await resetDerivedEntityDataInner(db, logger);
 
-    const replay = await replaySourceFacts(db, logger, { triggeredByUserId: deps.triggeredByUserId });
+    const replay = await replaySourceFacts(db, logger);
 
     if (deps.skipEnrichment) {
       return {
